Refactor to use a scene-model; implement a main menu system
Menu allows you to choose whether to quit or create a new world.
Noit yet productionised as the menu is created from scratch and
still needs to be generalised so it can be used in other
contexts.

For now, hitting ESC on the world map will take you back to the
main menu and allow you to create another world.